astar

+++ Path Planning/dijkstra.cpp
@@ -10,8 +10,8 @@
 
 using namespace std;
 
-stack<int> Dijkstra::findPath(int init, int goal, int cnt, vector<pair<int, float>> graph[]) {
-    priority_queue<pair<float, int>, vector<pair<float, int>>, greater<pair<float, int>>> pq;
+stack<int> Dijkstra::findPath(int init, int goal, int cnt, vector<pair<int, double>> graph[]) {
+    priority_queue<pair<double, int>, vector<pair<double, int>>, greater<pair<double, int>>> pq;
     vector<double> dist(cnt);
     vector<int> prev(cnt);
     stack<int> path;
@@ -19,7 +19,7 @@ stack<int> Dijkstra::findPath(int init, int goal, int cnt, vector<pair<int, floa
     vector<bool> done(cnt, false);
     
     for (int i = 0; i < cnt; i++) {
-        dist[i] = numeric_limits<float>::max();
+        dist[i] = numeric_limits<double>::max();
         prev[i] = -1;
     }
 
@@ -40,10 +40,10 @@ stack<int> Dijkstra::findPath(int init, int goal, int cnt, vector<pair<int, floa
             return path;
         }
 
-        vector<pair<int, float>>::iterator it;
+        vector<pair<int, double>>::iterator it;
         for (it = graph[u].begin(); it != graph[u].end(); it++) {
             int v = (*it).first;
-            float weight = (*it).second;
+            double weight = (*it).second;
             if (dist[v] > dist[u] + weight) {
                 dist[v] = dist[u] + weight;
                 prev[v] = u;
